feat(theme-builder): move CSSRuleObject to a dedicated ./core/css file
